+++ test/cookbooks/test/metadata.rb
@@ -1,0 +1,8 @@
+name 'test'
+maintainer 'Manuel Torrinha'
+maintainer_email 'manuel.torrinha@tecnico.ulisboa.pt'
+license 'Apache 2.0'
+description 'Cookbook for testing the shifter cookbook'
+version '0.1.0'
+
+depends 'shifter'