+++ test/cookbooks/test/recipes/shifter_compile.rb
@@ -1,0 +1,3 @@
+shifter_compile 'Compile and Install Shifter, and cleanup temp dir' do
+  action :install, :clean
+end