+++ test/cookbooks/test/recipes/shifter_install.rb
@@ -1,0 +1,3 @@
+shifter_install 'Install Shifter Runtime and make its binary available' do
+  action :install
+end