+++ test/cookbooks/test/recipes/shifterimg_install.rb
@@ -1,0 +1,3 @@
+shifter_install 'Install Shifter Image Gateway and make its binary available' do
+  action :install
+end